cleanup and structuring
