+++ game/save/load.go
@@ -1,0 +1,22 @@
+package save
+
+import (
+	"github.com/mlange-42/arche/ecs"
+	"github.com/mlange-42/tiny-world/game/comp"
+)
+
+func LoadWorld(world *ecs.World, path string) error {
+	_ = ecs.ComponentID[comp.Tile](world)
+	_ = ecs.ComponentID[comp.Terrain](world)
+	_ = ecs.ComponentID[comp.UpdateTick](world)
+	_ = ecs.ComponentID[comp.Consumption](world)
+	_ = ecs.ComponentID[comp.Production](world)
+	_ = ecs.ComponentID[comp.Warehouse](world)
+	_ = ecs.ComponentID[comp.BuildRadius](world)
+	_ = ecs.ComponentID[comp.Path](world)
+	_ = ecs.ComponentID[comp.Hauler](world)
+	_ = ecs.ComponentID[comp.HaulerSprite](world)
+	_ = ecs.ComponentID[comp.ProductionMarker](world)
+
+	return loadWorld(world, path)
+}
